ISD Week 3
In this week i wrote some code that added two numbers imputed from the user and added them together as integers. I then used python to answer a math question using different variables. as well as that i also ran a test on using two // instead of one/. Finally i answered some questions on what variable names can and cannot be, along with what the best option for a variable name would be.